+++ src/Tests/RefundOrderTest.php
@@ -5,6 +5,7 @@ namespace GingerPluginSdk\Tests;
 use GingerPluginSdk\Client;
 use GingerPluginSdk\Exceptions\InvalidOrderStatusException;
 use GingerPluginSdk\Properties\ClientOptions;
+use GingerPluginSdk\Properties\Amount;
 use PHPUnit\Framework\TestCase;
 
 class RefundOrderTest extends TestCase
@@ -29,5 +30,13 @@ class RefundOrderTest extends TestCase
         $order_data = OrderStub::getValidOrder();
         $order_obj = $this->client->sendOrder(order: $order_data);
         $this->client->refundOrder(order_id: $order_obj->getId()?->get());
+    }
+
+    public function test_invalid_partial_refund_uncompleted()
+    {
+        self::expectException(InvalidOrderStatusException::class);
+        $order_data = OrderStub::getValidOrder();
+        $order_obj = $this->client->sendOrder(order: $order_data);
+        $this->client->refundOrder(order_id: $order_obj->getId()?->get(),amount: new Amount(300));
     }
 }